move seeding script to /scripts and other misc cleanup
