+++ .env.example
@@ -53,17 +53,3 @@ MASTER_PASSWORD_HASH=""
 PROTECTED_SYMMETRIC_KEY=""
 GENERATED_RSA_KEY_PAIR_PUBLIC_KEY=""
 GENERATED_RSA_KEY_PAIR_PROTECTED_PRIVATE_KEY=""
-
-
-
-# openssl req -x509 -newkey rsa:4096 -keyout ssl.key -out ssl.crt -sha256 -days 1826 -nodes -subj "/CN=localhost/O=Bitwarden Autofill Testing" -addext "subjectAltName=DNS:localhost,DNS:bitwarden.test,IP:127.0.0.1"
-# # MAC OS:
-# sudo security add-trusted-cert -d -r trustRoot -k /Library/Keychains/System.keychain ssl.crt
-# NODE_EXTRA_CA_CERTS=ssl.crt NODE_OPTIONS=--use-openssl-ca bw login
-
-
-# openssl genrsa -out ca.key 2048
-# openssl req -new -x509 -days 1826 -key ca.key -subj "/C=US/ST=California/L=Santa Barbara/O=Bitwarden Autofill Testing/CN=Bitwarden Autofill Testing CA" -out ca.crt
-
-# openssl req -newkey rsa:2048 -nodes -keyout ssl.key -subj "/C=US/ST=California/L=Santa Barbara/O=Bitwarden Autofill Testing/CN=localhost,127.0.0.1,bitwarden.test" -out ssl.crt
-# openssl x509 -req -extfile <(printf "subjectAltName=DNS:localhost,127.0.0.1,bitwarden.test") -days 1826 -in ssl.crt -CA ca.crt -CAkey ssl.key -CAcreateserial -out ssl.crt
