+++ package.json
@@ -33,20 +33,20 @@
   "scripts": {
     "prepare": "husky install && npm run setup:clients",
     "pretest": "rimraf tests-out && rimraf screenshots && tsc --incremental -p tests/tsconfig.json",
-    "test:webserve": "cd clients/apps/web && npm run build:bit:watch",
-    "test:webserve:bitwarden": "cd clients/apps/web && ENV=cloud npm run build:oss:watch",
-    "test:autofill": "npm run pretest && playwright test --reporter=line",
-    "test:autofill:debug": "npm run pretest && playwright test --debug",
+    "build:clients": "cd clients/apps/browser && rimraf build && npm run build",
+    "build:clients:autofill": "cd clients/apps/browser && rimraf build && cross-env AUTOFILL_VERSION=2 npm run build",
+    "build:clients:autofill:mv3": "cd clients/apps/browser && rimraf build && cross-env AUTOFILL_VERSION=2 MANIFEST_VERSION=3 npm run build",
+    "build:clients:mv3": "cd clients/apps/browser && rimraf build && npm run build:mv3",
+    "build:clients:prod": "cd clients/apps/browser && rimraf build && npm run build:prod",
+    "seed:vault:account": "./scripts/create-account.sh",
+    "seed:vault:ciphers": "./scripts/cli-serve-helper.sh && ts-node ./scripts/vault-seeder/vault-seeder.ts && kill $(ps -e | grep 'bw serve' | grep -v 'vault-seeder' | awk '{print $1}')",
+    "seed:vault:ciphers:refresh": "./scripts/cli-serve-helper.sh && REFRESH=true ts-node ./scripts/vault-seeder/vault-seeder.ts && kill $(ps -e | grep 'bw serve' | grep -v 'vault-seeder' | awk '{print $1}')",
     "setup:clients": "rimraf clients && git clone https://github.com/bitwarden/clients.git clients && cd clients && npm ci",
     "setup:ssl": "./scripts/generate-certs.sh",
     "setup:vault": "npm run seed:vault:account && npm run seed:vault:ciphers",
-    "seed:vault:account": "./scripts/create-account.sh",
-    "seed:vault:ciphers": "./vault-seeder/bw-serve-helper.sh && REFRESH=true ts-node ./vault-seeder/vault-seeder.ts && kill $(ps -e | grep 'bw serve' | grep -v 'vault-seeder' | awk '{print $1}')",
-    "seed:vault:ciphers:refresh": "./vault-seeder/bw-serve-helper.sh && REFRESH=true ts-node ./vault-seeder/vault-seeder.ts && kill $(ps -e | grep 'bw serve' | grep -v 'vault-seeder' | awk '{print $1}')",
-    "build:clients": "cd clients/apps/browser && rimraf build && npm run build",
-    "build:clients:prod": "cd clients/apps/browser && rimraf build && npm run build:prod",
-    "build:clients:mv3": "cd clients/apps/browser && rimraf build && npm run build:mv3",
-    "build:clients:autofill": "cd clients/apps/browser && rimraf build && cross-env AUTOFILL_VERSION=2 npm run build",
-    "build:clients:autofill:mv3": "cd clients/apps/browser && rimraf build && cross-env AUTOFILL_VERSION=2 MANIFEST_VERSION=3 npm run build"
+    "test:autofill": "npm run pretest && playwright test --reporter=line",
+    "test:autofill:debug": "npm run pretest && playwright test --debug",
+    "test:webserve": "cd clients/apps/web && npm run build:bit:watch",
+    "test:webserve:bitwarden": "cd clients/apps/web && ENV=cloud npm run build:oss:watch"
   }
 }
