+++ tests/fixtures.ts
@@ -32,7 +32,6 @@ export const test = base.extend<{
       },
     });
     await use(context);
-    // await context.close();
   },
   extensionId: async ({ context }, use) => {
     let background;
